// src/worksheet/enums/worksheet.enums.ts
export enum enumOption {
  yes = "yes",
  no = "no",
}

export enum usedFor {
  exam = "exam",
  practice = "practice",
  test = "test",
}

export enum visibility {
  private = "private",
  public = "public",
}

export enum source {
  diksha = "diksha",
  khanAcademy = "khanAcademy",
}

export enum subjectList {
  hindi = "hindi",
  mathematics = "mathematics",
  bengali = "bengali",
  marathi = "marathi",
  english = "english",
  balbharatikannada = "balbharatikannada",
  cocurricular = "cocurricular",
  khelokarusikhu = "khelokarusikhu",
  tamil = "tamil",
  marathishikshaksanhita = "marathishikshaksanhita",
  khelukarushiku = "khelukarushiku",
  playdolearn = "playdolearn",
  telugu = "telugu",
  sindhi = "sindhi",
  gujarati = "gujarati",
  urdu = "urdu",
  khelekaresikhe = "khelekaresikhe",
  kannada = "kannada",
  geography = "geography",
  evs = "evs",
  evspart1 = "evspart1",
  evspart2 = "evspart2",
  scholarshipmarathi = "scholarshipmarathi",
  scholarshipintelligencetest = "scholarshipintelligencetest",
  scholarshipenglish = "scholarshipenglish",
  scholarshipmathematics = "scholarshipmathematics",
  historyandcivics = "historyandcivics",
  sugambharatikannada = "sugambharatikannada",
  science = "science",
  sanskrit = "sanskrit",
  history = "history",
  arabicbidaytularebia = "arabicbidaytularebia",
  defencestudies = "defencestudies",
  kumarbharatikannada = "kumarbharatikannada",
  historyandpoliticalscience = "historyandpoliticalscience",
  selfdevelopment = "selfdevelopment",
  spanish = "spanish",
  gulhaafarsi = "gulhaafarsi",
  secretarialpracticesp = "secretarialpracticesp",
}

export enum classList {
  class1 = "class 1",
  class2 = " class 2",
  class3 = "class 3",
  class4 = " class 4",
  class5 = " class 5",
  class6 = "class 6",
  class7 = "class 7",
  class8 = "class 8",
  class9 = "class 9",
  class10 = "class 10",
  class11 = "class 11",
  class12 = "class 12",
}